requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
av==11.0.0
beautifulsoup4==4.12.3
bert-score==0.3.13
black==24.4.2
bypy==1.8.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
coverage==7.5.1
cryptography==42.0.7
ctranslate2==4.1.0
cycler==0.12.1
datasets==2.19.0
debugpy==1.8.1
decorator==4.4.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
einops==0.7.0
evaluate==0.4.1
executing==2.0.1
faster-whisper==1.0.1
filelock==3.13.4
flake8==7.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
hanziconv==0.3.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
jellyfish==1.0.3
jieba==0.42.1
Jinja2==3.1.3
jiwer==3.0.4
joblib==1.4.0
jsonlines==4.0.0
julius==0.2.7
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==25.2.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning==2.2.2
lightning-utilities==0.11.2
llvmlite==0.42.0
lxml==5.2.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mne==1.6.1
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
nlpaug==1.1.11
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.17.3
openai==1.23.1
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pandas-jsonlines==1.0
parso==0.8.4
pathspec==0.12.1
peft==0.10.0
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.5.0
pooch==1.8.1
portalocker==2.8.2
primePy==1.3
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.1
rapidfuzz==3.9.0
readme_renderer==43.0
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.18.0
rfc3986==2.0.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.3
segtok==1.5.11
semver==3.0.2
sentencepiece==0.2.0
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.3.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.0+cu121
torchmetrics==1.3.2
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
twine==5.0.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
wheel==0.41.2
whisper==1.1.10
whisperx==3.1.3
xxhash==3.4.1
yake==0.4.8
yapf==0.40.2
yarl==1.9.4
zhconv==1.4.3
zipp==3.18.1
